organise les définitions de steps pour avoir à la fois des définitions globales et locales
