+++ e2e/cypress/integration/common/selectors.js
@@ -1,5 +1,5 @@
 /**
- * Sélecteurs d'éléments dans la page
+ * Sélecteurs d'éléments communs à toutes les pages
  */
 
 export const Selectors = {
@@ -14,41 +14,6 @@ export const Selectors = {
   },
   home: {
     selector: '[data-test=home]',
-  },
-  'formulaire de connexion': {
-    selector: '[data-test=SignInPage]',
-    children: {
-      email: 'input[name=email]',
-      mdp: 'input[name=password]',
-      Valider: 'button[type=submit]',
-      'Mot de passe oublié': '[data-test=forgotten-pwd]',
-    },
-  },
-  'demande de lien de réinitialisation du mdp': {
-    selector: '[data-test=PasswordRecovery]',
-    children: {
-      email: 'input[name=email]',
-      Valider: 'button[type=submit]',
-    },
-  },
-  'message lien envoyé': {
-    selector: '[data-test=PasswordRecoverySucceed]',
-    children: {
-      'Retour à la connexion': 'button',
-    },
-  },
-  'formulaire de réinitialisation du mdp': {
-    selector: '[data-test=ResetPassword]',
-    children: {
-      mdp: 'input[name=password]',
-      Valider: 'button[type=submit]',
-    },
-  },
-  'réinitialisation du mot de passe réussie': {
-    selector: '[data-test=ResetPasswordSucceed]',
-  },
-  'réinitialisation du mot de passe en erreur': {
-    selector: '[data-test=ResetPasswordFailed]',
   },
   'mes collectivités': {
     selector: '[data-test=CurrentUserCollectivites]',
